+++ dart/geobase/test/geodesy/spherical_ported_test.dart
@@ -28,7 +28,7 @@ void main() {
       const p2 = Geographic(lat: 48.857, lon: 2.351);
       const greenwich = Geographic(lat: 51.47788, lon: -0.00147);
 
-      expect(p1.latLonDms(), '52.2050°N 0.1190°E');
+      expect(p1.latLonDms(separator: ' '), '52.2050°N 0.1190°E');
 
       const latLonData = [
         ['52.205', '0.119'],
@@ -39,7 +39,7 @@ void main() {
       for (final latLon in latLonData) {
         expect(
           Geographic.parseDms(lat: latLon[0], lon: latLon[1]).latLonDms(),
-          '52.2050°N 0.1190°E',
+          '52.2050°N, 0.1190°E',
         );
       }
 
@@ -54,17 +54,17 @@ void main() {
       expect(sp1.finalBearingTo(p2).toStringAsFixed(1), '157.9');
       expect(
         sp1.midPointTo(p2).latLonDms(),
-        '50.5363°N 1.2746°E',
+        '50.5363°N, 1.2746°E',
       );
       expect(
         sp1.intermediatePointTo(p2, fraction: 0.25).latLonDms(),
-        '51.3721°N 0.7073°E',
+        '51.3721°N, 0.7073°E',
       );
       expect(
         greenwich.spherical
             .destinationPoint(distance: 7794, bearing: 300.7)
             .latLonDms(),
-        '51.5136°N 0.0983°W',
+        '51.5136°N, 0.0983°W',
       );
       expect(
         Geographic(lat: 51.8853, lon: 0.2545)
@@ -75,7 +75,7 @@ void main() {
               otherBearing: 32.435,
             )!
             .latLonDms(),
-        '50.9078°N 4.5084°E',
+        '50.9078°N, 4.5084°E',
       );
       expect(
         Geographic(lat: 53.2611, lon: -0.7972)
@@ -107,11 +107,11 @@ void main() {
       expect(rl3.finalBearingTo(p4).toStringAsFixed(1), '116.7');
       expect(
         rl3.destinationPoint(distance: 40300, bearing: 116.7).latLonDms(),
-        '50.9642°N 1.8530°E',
+        '50.9642°N, 1.8530°E',
       );
       expect(
         rl3.midPointTo(p4).latLonDms(),
-        '51.0455°N 1.5957°E',
+        '51.0455°N, 1.5957°E',
       );
 
       const polygon = [
@@ -129,15 +129,15 @@ void main() {
         Geographic(lat: 52.205, lon: 0.119),
       );
 
-      expect(greenwich.latLonDms(), '51.4779°N 0.0015°W');
+      expect(greenwich.latLonDms(separator: ' '), '51.4779°N 0.0015°W');
       expect(
         greenwich.latLonDms(
           format: Dms(type: DmsType.degMinSec, decimals: 2),
         ),
-        '51°28′40.37″N 0°00′05.29″W',
+        '51°28′40.37″N, 0°00′05.29″W',
       );
       expect(
-        greenwich.latLonDms(format: Dms.numeric()).split(' '),
+        greenwich.latLonDms(format: Dms.numeric(), separator: ' ').split(' '),
         ['51.4779', '-0.0015'],
       );
     });
@@ -162,15 +162,15 @@ void main() {
       expect(parisSp.initialBearingTo(cambg).toStringAsFixed(1), '337.9');
       expect(
         cambgSp.midPointTo(paris).latLonDms(),
-        '50.5363°N 1.2746°E',
+        '50.5363°N, 1.2746°E',
       );
       expect(
         cambgSp.intermediatePointTo(paris, fraction: 0.25).latLonDms(),
-        '51.3721°N 0.7073°E',
+        '51.3721°N, 0.7073°E',
       );
       expect(
         cambgSp.intermediatePointTo(cambg, fraction: 0.25).latLonDms(),
-        '52.2050°N 0.1190°E',
+        '52.2050°N, 0.1190°E',
       );
 
       final greenwichSp = Geographic(lat: 51.47788, lon: -0.00147).spherical;
@@ -178,7 +178,7 @@ void main() {
       const brng = 300.7;
       expect(
         greenwichSp.destinationPoint(distance: dist, bearing: brng).latLonDms(),
-        '51.5136°N 0.0983°W',
+        '51.5136°N, 0.0983°W',
       );
     });
   });
@@ -197,7 +197,7 @@ void main() {
         p01.spherical
             .intersectionWith(bearing: N, other: p10, otherBearing: E)!
             .latLonDms(),
-        '0.9998°N 1.0000°E',
+        '0.9998°N, 1.0000°E',
       );
     });
     test('toward 1,1 E,N nearest', () {
@@ -205,7 +205,7 @@ void main() {
         p10.spherical
             .intersectionWith(bearing: E, other: p01, otherBearing: N)!
             .latLonDms(),
-        '0.9998°N 1.0000°E',
+        '0.9998°N, 1.0000°E',
       );
     });
     test('toward 1,1 N,E antipodal', () {
@@ -243,7 +243,7 @@ void main() {
         p01.spherical
             .intersectionWith(bearing: S, other: p10, otherBearing: W)!
             .latLonDms(),
-        '0.9998°S 179.0000°W',
+        '0.9998°S, 179.0000°W',
       );
     });
     test('away 1,1 W,S antipodal', () {
@@ -251,7 +251,7 @@ void main() {
         p10.spherical
             .intersectionWith(bearing: W, other: p01, otherBearing: S)!
             .latLonDms(),
-        '0.9998°S 179.0000°W',
+        '0.9998°S, 179.0000°W',
       );
     });
     test('1E/90E N,E antipodal', () {
@@ -266,7 +266,7 @@ void main() {
         p01.spherical
             .intersectionWith(bearing: N, other: p1_92, otherBearing: E)!
             .latLonDms(),
-        '0.0175°N 179.0000°W',
+        '0.0175°N, 179.0000°W',
       );
     });
     test('coincident', () {
@@ -274,7 +274,7 @@ void main() {
         p11.spherical
             .intersectionWith(bearing: N, other: p11, otherBearing: E)!
             .latLonDms(),
-        '1.0000°N 1.0000°E',
+        '1.0000°N, 1.0000°E',
       );
     });
 
@@ -290,7 +290,7 @@ void main() {
               otherBearing: 32.435,
             )!
             .latLonDms(),
-        '50.9078°N 4.5084°E',
+        '50.9078°N, 4.5084°E',
       );
     });
     test('rounding errors', () {
@@ -303,7 +303,7 @@ void main() {
               otherBearing: 60,
             )!
             .latLonDms(),
-        '50.4921°N 1.3612°E',
+        '50.4921°N, 1.3612°E',
       );
     });
 
@@ -328,7 +328,7 @@ void main() {
               otherBearing: 180,
             )!
             .toDmsLatLon(),
-        '90.0000°S 163.9902°W',
+        '90.0000°S, 163.9902°W',
       );
     });
   */
@@ -469,13 +469,13 @@ void main() {
     test('parallels 1', () {
       expect(
         Geographic(lat: 30, lon: parallels![0]).latLonDms(format: dms),
-        '30°00′00″N 9°35′39″E',
+        '30°00′00″N, 9°35′39″E',
       );
     });
     test('parallels 2', () {
       expect(
         Geographic(lat: 30, lon: parallels![1]).latLonDms(format: dms),
-        '30°00′00″N 170°24′21″E',
+        '30°00′00″N, 170°24′21″E',
       );
     });
     test('parallels -', () {
@@ -656,7 +656,7 @@ void main() {
         lax.spherical
             .intermediatePointTo(jfk, fraction: 100.0 / 2144.0)
             .latLonDms(format: Dms(type: DmsType.degMin, decimals: 0)),
-        '34°37′N 116°33′W',
+        '34°37′N, 116°33′W',
       );
     });
 
@@ -683,7 +683,7 @@ void main() {
         lax.spherical
             .intermediatePointTo(jfk, fraction: 0.4)
             .latLonDms(format: Dms(type: DmsType.degMin, decimals: 3)),
-        '38°40.167′N 101°37.570′W',
+        '38°40.167′N, 101°37.570′W',
       );
     });
 
@@ -695,7 +695,7 @@ void main() {
         reo.spherical
             .intersectionWith(bearing: 51, other: bke, otherBearing: 137)!
             .latLonDms(format: Dms(decimals: 3)),
-        '43.572°N 116.189°W',
+        '43.572°N, 116.189°W',
       );
     });
     test('', () {});
@@ -762,14 +762,14 @@ void main() {
     test('dest’n', () {
       expect(
         dov.rhumb.destinationPoint(distance: 40310, bearing: 116.7).latLonDms(),
-        '50.9641°N 1.8531°E',
+        '50.9641°N, 1.8531°E',
       );
       expect(
         Geographic(lat: 1, lon: 1)
             .rhumb
             .destinationPoint(distance: 111178, bearing: 90)
             .latLonDms(),
-        '1.0000°N 2.0000°E',
+        '1.0000°N, 2.0000°E',
       );
     });
     test('dest’n dateline', () {
@@ -778,20 +778,20 @@ void main() {
             .rhumb
             .destinationPoint(distance: 222356, bearing: 90)
             .latLonDms(),
-        '1.0000°N 179.0000°W',
+        '1.0000°N, 179.0000°W',
       );
       expect(
         Geographic(lat: 1, lon: -179)
             .rhumb
             .destinationPoint(distance: 222356, bearing: 270)
             .latLonDms(),
-        '1.0000°N 179.0000°E',
+        '1.0000°N, 179.0000°E',
       );
     });
     test('midpoint', () {
       expect(
         dov.rhumb.midPointTo(cal).latLonDms(),
-        '51.0455°N 1.5957°E',
+        '51.0455°N, 1.5957°E',
       );
     });
     test('midpoint dateline', () {
@@ -800,7 +800,7 @@ void main() {
             .rhumb
             .midPointTo(Geographic(lat: 1, lon: 178))
             .latLonDms(),
-        '1.0000°N 179.5000°E',
+        '1.0000°N, 179.5000°E',
       );
     });